requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.1.2
Django==1.11
django-common-helpers==0.9.1
django-cron==0.5.0
django-filter==1.1.0
django-simple-history==1.9.0
djangorestframework==3.4.0
djangorestframework-jwt==1.11.0
enum34==1.1.6
gunicorn==19.7.1
ipython==5.5.0
ipython-genutils==0.2.0
Markdown==2.6.9
olefile==0.44
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
pytz==2017.3
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
